feat(schema): add support for multi-value properties
BREAKING CHANGE: this is applying @tripodsan's changes from https://github.com/adobe/helix-index-pipelines/pull/38 to the Index Config schema. In particular it requires each property to have either one `value` or one `values` properties, but neve both. As this changes the behavior of `value`, it is a breaking change.

fixes #212