refactor(fstab): change NamedPairHandler to MountPointHandler
`NamedPairHandler` has been renamed to `MountPointHandler` and taken over the getter from `MountConfig`, so that no sanitizing and no modifications are needed anymore. The ability to override the type or set custom properties on a mount point is gone, but the config file looks much cleaner and simpler now.

alternative to (closes) #247