fix(redirects): handle special case of äӓ confusion
the blog is using a url with ä (cyrillic) instead of ӓ (latin) and redirect definitions do not take hold because when entering the cyrillic character in excel, it will autocorrect to the latin character, which does not match, and there is no adequate normalization available either. So the (narrow) workaround is to put the URL encoded variants (both, in both possible capitalizations) into the spreadsheet and extend the matching to URL encode both variants of ä and ӓ when encountered.